+++ usage-stats/lib/client-src/10-loader-open.js
@@ -1,0 +1,11 @@
+window.__ModuleLoader__.load({
+  id: "dsh-usage-stats",
+  factory: (require) => {
+    var module = { exports: {} };
+    var exports = module.exports;
+    Object.defineProperty(exports, Symbol.toStringTag, { value: "Module" });
+
+    var React = require("react");
+    var inject = ["slots"];
+
+    var API = "/api/usage-stats/summary";